+++ packages/modules/imodel-browser/jest.config.js
@@ -1,3 +1,7 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
 module.exports = {
   roots: ["<rootDir>/src"],
   transform: {
@@ -9,7 +13,7 @@ module.exports = {
   ],
   moduleNameMapper: {
     "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga|svg?.+)$":
-      "<rootDir>/src/tests/mocks/fileMock.js",
+      "<rootDir>/src/tests/mocks/fileMock.ts",
     "\\.(css|less|scss|sass)$": "identity-obj-proxy",
   },
   verbose: true,